+++ Source/DREAM3DLib/StatisticsFilters/FindSizes.cpp
@@ -164,7 +164,7 @@ void FindSizes::find_sizes()
   {
     graincounts[i] = 0.0f;
   }
-  for (int j = 0; j < totalPoints; j++)
+  for (size_t j = 0; j < totalPoints; j++)
   {
     int gnum = m_GrainIds[j];
     graincounts[gnum]++;
@@ -197,7 +197,7 @@ void FindSizes::find_sizes2D()
   {
       graincounts[i] = 0.0f;
   }
-  for (int j = 0; j < totalPoints; j++)
+  for (size_t j = 0; j < totalPoints; j++)
   {
     int gnum = m_GrainIds[j];
     graincounts[gnum]++;
